Create Global economic overview of 2020
The projects shows the economic overview from 2020/21. The data is from the COVID-19 dataset. I explored the basic socio-economic concept such as; GDP, Per Capita Income, Population density, reproduction rate and much more. The aim was to split these concepts by continents and location, then rank them to see which continent and location was dominant in which aspect. The visualization further helps to show this in simple but insightful form, the spread of population among continent, the distribution of global wealth, the countries rich in humans resource and so much more.